+++ src/main/java/roomescape/domain/reservationtime/ReservationTimeService.java
@@ -7,10 +7,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.reservationdate.ReservationDateRepository;
-import roomescape.domain.reservationtime.dto.CreateTimeRequest;
-import roomescape.domain.reservationtime.dto.CreateTimeResponse;
+import roomescape.domain.reservationtime.admin.CreateTimeRequest;
+import roomescape.domain.reservationtime.admin.CreateTimeResponse;
 import roomescape.domain.reservationtime.dto.ReservationTimeAvailabilityResponse;
-import roomescape.domain.reservationtime.dto.ReservationTimeResponse;
+import roomescape.domain.reservationtime.admin.ReservationTimeResponse;
 import roomescape.domain.theme.ThemeRepository;
 import roomescape.support.exception.ConflictException;
 import roomescape.support.exception.NotFoundException;
