+++ src/test/java/roomescape/domain/reservationtime/ReservationTimeServiceTest.java
@@ -12,10 +12,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservationdate.ReservationDate;
-import roomescape.domain.reservationtime.dto.CreateTimeRequest;
-import roomescape.domain.reservationtime.dto.CreateTimeResponse;
+import roomescape.domain.reservationtime.admin.CreateTimeRequest;
+import roomescape.domain.reservationtime.admin.CreateTimeResponse;
 import roomescape.domain.reservationtime.dto.ReservationTimeAvailabilityResponse;
-import roomescape.domain.reservationtime.dto.ReservationTimeResponse;
+import roomescape.domain.reservationtime.admin.ReservationTimeResponse;
 import roomescape.domain.theme.Theme;
 import roomescape.support.exception.NotFoundException;
 import roomescape.support.exception.RoomescapeException;
